+++ jive-android-core-entity/src/main/java/com/jivesoftware/android/mobile/sdk/entity/TokenEntity.java
@@ -20,5 +20,5 @@ public class TokenEntity {
     public String tokenType;
 
     @JsonProperty("expires_in")
-    public int expiresIn;
+    public long expiresIn;
 }
